Support U2F/WebAuthn keys USB detection in Linux
This commit updates marshallbrekka/go-u2fhost, imported by
the Okta and Google providers to support U2F/WebAuthn USB keys
as 2FA devices.

The update should help fix the problems reported in
https://github.com/Versent/saml2aws/issues/419 as they had their
origin in marshallbrekka/go-u2fhost not able to properly detect
these devices in Linux. That was fixed upstream with
https://github.com/marshallbrekka/go-u2fhost/pull/10 which also
explains what other changes were done in the hid libraries
used by go-u2fhost to make the detection work in Linux too.

Along the library change and other mod updates, this commit includes
a change to the makefile to support the required tag to use
hidraw instead of libusb when compiling linux (and so really apply
the fix when compiling) as well as related changes to make it easier
to build/install saml2aws in different platforms, including OS
detection to prevent macOS builds from building in non-Darwin
environments, which should tackle
https://github.com/Versent/saml2aws/issues/56.

Finally, I've also included some updates in the README
to clarify the changes in the build process and the additional
per OS targets.